third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/blob-popup.https-expected.txt

This is a testharness.js-based test.
FAIL Cross-Origin-Opener-Policy and a blob URL popup assert_equals: expected "" but got "65df0e4c-0b76-43b3-ab1d-a9d31fa4273d"
Harness: the test ran to completion.
